+++ core-ng/src/main/java/core/framework/impl/web/websocket/ChannelCallback.java
@@ -9,11 +9,11 @@ import org.slf4j.LoggerFactory;
  * @author neo
  */
 public class ChannelCallback implements WebSocketCallback<Void> {
+    static final ChannelCallback INSTANCE = new ChannelCallback();
     private final Logger logger = LoggerFactory.getLogger(ChannelCallback.class);
 
     @Override
     public void complete(WebSocketChannel channel, Void context) {
-
     }
 
     @Override