+++ core-ng/src/test/java/core/framework/impl/web/http/CIDRTest.java
@@ -11,7 +11,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 class CIDRTest {
     @Test
     void matchesAll() {
-        CIDR cidr = new CIDR("0.0.0.0/0");
+        var cidr = new CIDR("0.0.0.0/0");
 
         assertThat(cidr.matches(address("192.168.1.1"))).isTrue();
         assertThat(cidr.matches(address("127.0.0.1"))).isTrue();
@@ -27,7 +27,7 @@ class CIDRTest {
 
     @Test
     void matchesWithIPv4() {
-        CIDR cidr = new CIDR("192.168.1.0/24");
+        var cidr = new CIDR("192.168.1.0/24");
 
         assertThat(cidr.matches(address("192.168.1.1"))).isTrue();
         assertThat(cidr.matches(address("192.168.1.254"))).isTrue();
@@ -64,7 +64,7 @@ class CIDRTest {
 
     @Test
     void matchesWithIPv6() {
-        CIDR cidr = new CIDR("2001:0db8:85a3:0000:0000:8a2e:0370:0000/112");
+        var cidr = new CIDR("2001:0db8:85a3:0000:0000:8a2e:0370:0000/112");
 
         assertThat(cidr.matches(address("2001:0db8:85a3:0000:0000:8a2e:0370:7334"))).isTrue();
     }